post-publisher/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure-functions==1.20.0
azure-storage-blob==12.22.0
azure-cosmos==4.7.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
python-dotenv==1.0.1
requests==2.32.3
pillow==10.4.0
pytest==8.3.2